requirements.txt
flask==3.0.3
openai>=1.57.0
gspread==6.1.2
google-auth==2.29.0
pymupdf==1.24.10
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
